sqltojson/requirements.txt
pymysql
paho-mqtt
cryptography
